moved fixed-lenght tests to its own file
